Move personas CLI tests to integration dir
